fixed ButtonRound and SignClient
