+++ src/Components/Login/Signup.jsx
@@ -1,27 +1,8 @@
-import './Login.scss';
-import { useState } from 'react';
 import { Link } from 'react-router-dom';
-import { createUsuarios } from '../../Services/user';
+import './Login.scss';
+import ButtonRound from '../ButtonRound/ButtonRound';
 
-function Signup() {
-  const [form, setForm] = useState({});
-  const handleChange = (e) => {
-    const { name, value } = e.target;
-    setForm({
-      ...form,
-      [name]: value,
-    });
-  };
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-    // const datatosend = {
-    //   ...form, profile: 'imagen.png',
-    // };
-    await createUsuarios(form);
-    setForm({});
-    document.querySelector('form').reset();
-  };
-
+export default function Signup() {
   return (
     <div className="login">
       <div>
@@ -29,36 +10,31 @@ function Signup() {
       </div>
       <div className="principal">
         <Link to="/">
-          <img className="logo" src="images/logo/quick-call-logo--colored.svg" alt="" />
+          <img className="logo" src="images/logo/quick-call-logo.svg" alt="" />
         </Link>
         <div className="texto">
-          <span className="titulo_register">Registrarse</span>
-          <span className="texto_register">Crea tu nueva cuenta</span>
+          <span className="titulo_register"> Registrarse</span>
+          <span className="texto_register">Por favor escoge tu perfil</span>
         </div>
-        <div className="redes_sociales">
-          <img src="images/icons/whatsapp-logo.svg" alt="whatsapp" />
-          <img src="images/icons/facebook-icon.svg" alt="facebook" />
-          <img src="images/icons/twitter-icon.svg" alt="twitter" />
-          <img src="images/icons/linkedin-logo.svg" alt="linkedin" />
-        </div>
-        <form className="formulario" onSubmit={handleSubmit}>
-          <input name="name" placeholder="Ingresa tu nombre" type="text" onChange={handleChange} />
-          <input name="email" placeholder="Michelle@ejemplo.com" type="email" onChange={handleChange} />
-          <input name="password" placeholder="Ingresa tu contraseña" type="password" onChange={handleChange} />
-          <input name="password2" placeholder="Repite tu contraseña" type="password" onChange={handleChange} />
+        <form className="formulario">
+          <div className="boton11">
+            <Link to="/signup_client"><ButtonRound href="/signup_client">Cliente</ButtonRound></Link>
+            <Link to="/signup_professional"><ButtonRound>Profesional</ButtonRound></Link>
+          </div>
+          <div className="redes_sociales">
+            <img src="images/icons/whatsapp-logo.svg" alt="whatsapp" />
+            <img src="images/icons/facebook-icon.svg" alt="facebook" />
+            <img src="images/icons/twitter-icon.svg" alt="twitter" />
+            <img src="images/icons/linkedin-logo.svg" alt="linkedin" />
+          </div>
           <div className="footer1">
             <span className="footer11">
-              Ya tienes un cuenta?
-              <Link to="/login">Ingresa</Link>
+              ¿Ya tienes una cuenta?.
+              <Link to="/login">Ingresa aqui</Link>
             </span>
-            <button className="boton" type="submit">
-              Registrarse
-            </button>
           </div>
         </form>
       </div>
     </div>
   );
 }
-
-export default Signup;
